==> 1_gather_embeddings.ipynb <==
import os
os.getcwd()
from datetime import datetime
import pickle
import random
import re
import numpy as np
random.seed(21)
import torch
from transformers import DistilBertForMaskedLM, DistilBertConfig, AutoTokenizer

# TEXT DATA

### Sentences Containing Country Cases

year_list = [
    '1900_1909',
    '1910_1919',
    '1920_1929',
    '1930_1939',
    '1940_1949',
    '1950_1959',
    '1960_1969',
    '1970_1979',
    '1980_1989',
    '1990_1999',
    '2000_2009',
    '2010_2020'
]

country_list = ['china', 'north korea', 'south korea', 'canada', 'united kingdom', 'germany']

country_sentence_dict = {}

country_mask_dict = {}

def clean_sents(sents, term):
    terms = term.split(" ")
    sents = [sent.strip() for sent in sents]
    mask = [all(item in sent.split(" ") for item in terms) for sent in sents]
    sents = np.array(sents)
    result = sents[mask]
    
    return result.tolist(), mask

# mask isn't necessarily needed anymore but might be useful somehow
def get_decade_sentences(years, term):
    year_dict = {}
    mask_dict = {}
    for year in years:
        with open('./crs_corpus/decade_' + year + '.txt') as f:
            sents = [line for line in f.readlines() if term in line]
            sents, mask = clean_sents(sents, term)
            year_dict[year] = sents
            mask_dict[year] = mask
            
    return year_dict, mask_dict

#### "China"

sents_china, mask_dict = get_decade_sentences(year_list, 'china')
country_sentence_dict['china'] = sents_china
country_mask_dict['china'] = mask_dict

#### "North Korea"

sents_north_korea, mask_dict = get_decade_sentences(year_list, 'north korea')
country_sentence_dict['north korea'] = sents_north_korea
country_mask_dict['north korea'] = mask_dict

#### "South Korea"

sents_south_korea, mask_dict = get_decade_sentences(year_list, 'south korea')
country_sentence_dict['south korea'] = sents_south_korea
country_mask_dict['south korea'] = mask_dict

#### "Canada"

sents_canada, mask_dict = get_decade_sentences(year_list, 'canada')
country_sentence_dict['canada'] = sents_canada
country_mask_dict['canada'] = mask_dict

#### "United Kingdom"

sents_united_kingdom, mask_dict = get_decade_sentences(year_list, 'united kingdom')
country_sentence_dict['united kingdom'] = sents_united_kingdom
country_mask_dict['united kingdom'] = mask_dict

#### "Germany"

sents_germany, mask_dict = get_decade_sentences(year_list, 'germany')
country_sentence_dict['germany'] = sents_germany
country_mask_dict['germany'] = mask_dict

# SAVE COUNTRY MASKS

with open('./crs_sents/country_sent_masks.pkl', 'wb') as handle:
    pickle.dump(country_mask_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD COUNTRY MASKS

#with open('./crs_sents/country_sent_masks.pkl', 'rb') as handle:
#    country_mask_dict = pickle.load(handle)

# SAVE COUNTRY SENTENCES

with open('./crs_sents/country_sents.pkl', 'wb') as handle:
    pickle.dump(country_sentence_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD COUNTRY SENTENCES

#with open('./crs_sents/country_sents.pkl', 'rb') as handle:
#    country_sentence_dict = pickle.load(handle)

concept_list = ['authoritarianism',
                'autocracy',
                'autocratic',
                'democracy',
                'dictator',
                'dictatorship']

### Sentences Containing Concepts

Same thing for the following concepts:

concept_sentence_dict = {}

concept_mask_dict = {}

#### "authoritarianism"

sents_authoritarianism, mask_dict = get_decade_sentences(year_list, 'authoritarianism')
concept_sentence_dict['authoritarianism'] = sents_authoritarianism
concept_mask_dict['authoritarianism'] = mask_dict

#### "autocracy"

sents_autocracy, mask_dict = get_decade_sentences(year_list, 'autocracy')
concept_sentence_dict['autocracy'] = sents_autocracy
concept_mask_dict['autocracy'] = mask_dict

#### "autocratic"

sents_autocratic, mask_dict = get_decade_sentences(year_list, 'autocratic')
concept_sentence_dict['autocratic'] = sents_autocratic
concept_mask_dict['autocratic'] = mask_dict

#### "democracy"

sents_democracy, mask_dict = get_decade_sentences(year_list, 'democracy')
concept_sentence_dict['democracy'] = sents_democracy
concept_mask_dict['democracy'] = mask_dict

#### "dictator"

sents_dictator, mask_dict = get_decade_sentences(year_list, 'dictator')
concept_sentence_dict['dictator'] = sents_dictator
concept_mask_dict['dictator'] = mask_dict

#### "dictatorship"

sents_dictatorship, mask_dict = get_decade_sentences(year_list, 'dictatorship')
concept_sentence_dict['dictatorship'] = sents_dictatorship
concept_mask_dict['dictatorship'] = mask_dict

# SAVE CONCEPT SENTENCE MASKS

with open('./crs_sents/concept_sent_masks.pkl', 'wb') as handle:
    pickle.dump(concept_mask_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD CONCEPT SENTENCE MASKS

#with open('./crs_sents/concept_sent_masks.pkl', 'rb') as handle:
#    concept_mask_dict = pickle.load(handle)

# SAVE CONCEPT SENTENCES

with open('./crs_sents/concept_sents.pkl', 'wb') as handle:
    pickle.dump(concept_sentence_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD CONCEPT SENTENCES

#with open('./crs_sents/concept_sents.pkl', 'rb') as handle:
#    concept_sentence_dict = pickle.load(handle)

# MODELS & FUNCTIONS

### Custom Functions

layers = [-3, -2, -1]


def get_word_idx(sent: str, word: str):
    if ' ' in word:
        word = word.split(' ')
        sent = sent.split(' ')
        idx_list = [sent.index(x) for x in word]
        return idx_list
    else:
        return sent.split(" ").index(word)

def get_hidden_states(encoded, token_ids_word, model, layers):
    """Push input IDs through model. Stack and sum `layers` (last four by default).
       Select only those subword token outputs that belong to our word of interest
       and average them."""
    with torch.no_grad():
        output = model(**encoded)

    # Get all hidden states
    states = output.hidden_states
    # Stack and sum all requested layers
    output = torch.stack([states[i] for i in layers]).sum(0).squeeze()
    # Only select the tokens that constitute the requested word
    word_tokens_output = output[token_ids_word]

    return word_tokens_output.mean(dim=0)


def get_word_vector(sent, idx, tokenizer, model, layers):
    """Get a word vector by first tokenizing the input sentence, getting all token idxs
       that make up the word of interest, and then `get_hidden_states`."""
    encoded = tokenizer.encode_plus(sent, truncation=True, return_tensors="pt")
    # get all token idxs that belong to the word of interest
    
    if isinstance(idx, list):

        token_ids_word = np.where(np.isin(np.array(encoded.word_ids()), (idx)))
    else:
        token_ids_word = np.where(np.array(encoded.word_ids()) == idx)

    return get_hidden_states(encoded, token_ids_word, model, layers)

def load_decade_model(year):
    config = DistilBertConfig.from_json_file('./crs_models/' + year + '/config.json')
    config.output_hidden_states = True
    model = DistilBertForMaskedLM(config)
    model.load_state_dict(torch.load('./crs_models/' + year + '/pytorch_model.bin', map_location = torch.device("cuda")))
    model.eval()
    tokenizer = AutoTokenizer.from_pretrained('./crs_models/' + year + '/')
    return model, tokenizer

def extract_embedding(sent, term, tokenizer, model, layers):
    idx = get_word_idx(sent, term)
    term_tensor = get_word_vector(sent, idx, tokenizer, model, layers)
    return term_tensor

# No longer needed but useful for the future
def ensure_spaces(sent, term):
    sent = re.sub(rf'(?<![ ])({term})', r' \1', sent)
    sent = re.sub(rf'({term})(?![ ])', r'\1 ', sent)
    
    return sent

def get_embeddings_from_sents(year_list, term, sentence_dict, layers, random_sample = True):
    
    vector_dict = {}
    
    if random_sample == True:
        shortest_decade = min([len(sentence_dict[term][x]) for x in sentence_dict[term]])
    print("Starting " + term + " at " + str(datetime.now()))    
    for year in year_list:
        vector_list = []
        
        sents = sentence_dict[term][year]
        
        if len(sents) == 0:
            vector_dict[year] = None
        else:
        
            year_model, year_tokenizer = load_decade_model(year)


            if random_sample == True:
                sents = random.sample(sents,shortest_decade)

            for sent in sents:
                vector = extract_embedding(sent, term, year_tokenizer, year_model, layers)
                vector_list.append(vector)

            vector_list = [v.detach().cpu().numpy() for v in vector_list]    
            vector_dict[year] = vector_list
        print("Finished " + year + " at " + str(datetime.now()))
    return vector_dict

# CONTEXTUAL EMBEDDINGS FOR COUNTRIES

### China

country_embeddings = {}

te_china = get_embeddings_from_sents(year_list, 'china', country_sentence_dict, layers, random_sample = False)

country_embeddings['china'] = te_china

### North Korea

te_north_korea = get_embeddings_from_sents(year_list, 'north korea', country_sentence_dict, layers, random_sample = False)

country_embeddings['north korea'] = te_north_korea

### South Korea

te_south_korea = get_embeddings_from_sents(year_list, 'south korea', country_sentence_dict, layers, random_sample = False)

country_embeddings['south korea'] = te_south_korea

### Canada

te_canada = get_embeddings_from_sents(year_list, 'canada', country_sentence_dict, layers, random_sample = False)

country_embeddings['canada'] = te_canada

### United Kingdom

te_united_kingdom = get_embeddings_from_sents(year_list, 'united kingdom', country_sentence_dict, layers, random_sample = False)

country_embeddings['united kingdom'] = te_united_kingdom

### Germany

te_germany = get_embeddings_from_sents(year_list, 'germany', country_sentence_dict, layers, random_sample = False)

country_embeddings['germany'] = te_germany

# Not every country appears in every decade

for country in ['canada', 'united kingdom', 'germany', 'north korea', 'south korea']:
    for year in year_list:
        if country_embeddings[country][year] != None:
            print(country)
            print(year)
            print(len(country_embeddings[country][year]))

# SAVE COUNTRY EMBEDDINGS

with open('./crs_embeds/country_embeds.pkl', 'wb') as handle:
    pickle.dump(country_embeddings, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD COUNTRY EMBEDDINGS

#with open('./crs_embeds/country_embeds.pkl', 'rb') as handle:
#    country_embeddings = pickle.load(handle)

# (CONTEXTUAL) EMBEDDINGS FOR THEORETICAL CONCEPTS

Same thing for the following concepts listed above.

concept_embeddings = {}

#### "authoritarianism"

te_authoritarianism = get_embeddings_from_sents(year_list, 'authoritarianism', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['authoritarianism'] = te_authoritarianism

#### "autocracy"

te_autocracy = get_embeddings_from_sents(year_list, 'autocracy', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['autocracy'] = te_autocracy

#### "autocratic"

te_autocratic = get_embeddings_from_sents(year_list, 'autocratic', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['autocratic'] = te_autocratic

#### "democracy"

te_democracy = get_embeddings_from_sents(year_list, 'democracy', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['democracy'] = te_democracy

#### "dictator"

te_dictator = get_embeddings_from_sents(year_list, 'dictator', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['dictator'] = te_dictator

#### "dictatorship"

te_dictatorship = get_embeddings_from_sents(year_list, 'dictatorship', concept_sentence_dict, layers, random_sample = False)

concept_embeddings['dictatorship'] = te_dictatorship

# SAVE CONCEPT EMBEDDINGS

with open('./crs_embeds/concept_embeds.pkl', 'wb') as handle:
    pickle.dump(concept_embeddings, handle, protocol=pickle.HIGHEST_PROTOCOL)

# LOAD CONCEPT EMBEDDINGS

#with open('./crs_embeds/concept_embeds.pkl', 'rb') as handle:
#    concept_embeddings = pickle.load(handle)
